Recommend the Korifi installer job instead of `deploy-on-kind`
`deploy-on-kind` has been always meant to be used for dev purposes to
install Korifi from source code. This implies that the dev team makes no
guarantees that the script will always work, even exist in future.
Furthermore, building Korifi from source code requires that users need
to have the Korifi dev toolchain installed and that might be a bit
awkward for some users.

Instead, the tutorial recommends using the official installer job that
is maintained and documented by the maintaners.